+++ src/pydocformatter/rules/definitions/PDF/PDF503_extraneous_return_documentation.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
+import pydocformatter.rules.definition_helpers.docstring_conventions as docstring_conventions
 import pydocformatter.rules.definition_helpers.value_documentation as value_documentation
 import pydocformatter.rules.definitions.PDF.PDF as PDF_definition
 import pydocformatter.rules.registration as rule_registration
+from pydocformatter.cli.settings_check import DocstringConvention
 from pydocformatter.rules.codes import RuleCode
 from pydocformatter.rules.definition import RuleBase, RuleContext
 from pydocformatter.rules.definitions.PDF.PDF import PDF
-from pydocformatter.rules.models import FixAvailability, RuleFinding, RuleMetadata
+from pydocformatter.rules.models import FixAvailability, RuleFinding, RuleMetadata, RuleSettingEffect, RuleSettingEffects, RuleSettingEffectValues
 
 
 @rule_registration.register_rule_to(PDF)
@@ -17,7 +19,16 @@ class PDF503ExtraneousReturnDocumentation(RuleBase):
         message="Docstring has a return section for a function that does not return",
         fix_availability=FixAvailability.NEVER,
         stable_since="1.0.0",
-        setting_effects=(),
+        setting_effects=(
+            RuleSettingEffects(
+                setting="docstring_convention",
+                effects=(
+                    RuleSettingEffectValues(
+                        effect=RuleSettingEffect.IGNORED, values=docstring_conventions.ignored_conventions_except(DocstringConvention.GOOGLE, DocstringConvention.NUMPY, DocstringConvention.REST)
+                    ),
+                ),
+            ),
+        ),
         incompatible_with=(),
     )